added an defaultfunction for onLoad props in MovieItem
