+++ src/pages/Main.js
@@ -1,31 +1,59 @@
-import { Fragment, useEffect, useState } from "react";
+import { Fragment, useEffect, useState, useRef } from "react";
 import { useSelector } from "react-redux";
 import useHttp from "../hooks/useHttp";
 import { getAllPopular } from "../lib/api";
 
 import Loader from "../components/layout/Loader/Loader";
-import MovieList from "../components/movies/MovieList";
+import MovieItem from "../components/movies/MovieItem";
+import GridContainer from "../components/layout/Grid/GridContainer";
 
 const Main = () => {
   const { sendRequest, data, error, status } = useHttp(getAllPopular);
-  const [isLoading, setIsLoading] = useState(true);
-
-  const LoadingHandler = (value) => {
-    setIsLoading(value);
-  };
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     sendRequest();
   }, [sendRequest]);
 
+  const favouriteList = useSelector((state) => state.favourite.list);
+  const watchLaterList = useSelector((state) => state.watchLater.list);
+
+  const counter = useRef(0);
+
+  const itemIsLoaded = () => {
+    counter.current += 1;
+    console.log(counter);
+
+    if (counter.current === data.results.length - 1) {
+      setLoading(false);
+    }
+  };
+
   return (
-    <Fragment>
+    <>
       <h1>Popular Movie</h1>
-      {isLoading && <Loader />}
-      {!error && status === "completed" && (
-        <MovieList results={data.results} loading={LoadingHandler} />
-      )}
-    </Fragment>
+      {loading ? <Loader /> : ""}
+      <GridContainer>
+        {!error &&
+          status === "completed" &&
+          data.results.map((movie) => {
+            const favourited = favouriteList.find((x) => x.id === movie.id);
+            const watchLater = watchLaterList.find((x) => x.id === movie.id);
+
+            return (
+              <MovieItem
+                key={movie.id}
+                id={movie.id}
+                title={movie.title}
+                posterPath={movie.poster_path}
+                favourited={favourited ? true : false}
+                watchLater={watchLater ? true : false}
+                onLoad={itemIsLoaded}
+              />
+            );
+          })}
+      </GridContainer>
+    </>
   );
 };
